Fold announced version into version endpoint
