+++ __tests__/hooks/useVersion.test.tsx
@@ -5,6 +5,8 @@ describe("useIsVersionStale", () => {
     jest.useFakeTimers();
     globalThis.fetch = jest.fn();
     globalThis.history.replaceState(null, "", "/");
+    const { publicEnv } = require("@/config/env");
+    publicEnv.ANNOUNCED_VERSION_ENDPOINT = undefined;
   });
 
   afterEach(() => {
@@ -21,29 +23,75 @@ describe("useIsVersionStale", () => {
     const { publicEnv } = require("@/config/env");
     publicEnv.VERSION = "1.0.0";
     (globalThis.fetch as jest.Mock).mockResolvedValue({
-      json: async () => ({ version: "1.0.0" }),
+      json: async () => ({
+        announced_version: null,
+        stale: false,
+        version: "1.0.0",
+      }),
     });
     const { findByText } = render(<TestComponent interval={1000} />);
     await act(async () => {
       jest.runOnlyPendingTimers();
     });
-    expect(globalThis.fetch).toHaveBeenCalledWith("/api/announced-version", {
+    expect(globalThis.fetch).toHaveBeenCalledWith("/api/version", {
       cache: "no-store",
     });
     expect(await findByText("fresh")).toBeInTheDocument();
   });
 
-  it("shows stale when versions differ", async () => {
+  it("shows stale when versions differ and no announcement endpoint is configured", async () => {
     const { publicEnv } = require("@/config/env");
     publicEnv.VERSION = "1.0.0";
     (globalThis.fetch as jest.Mock).mockResolvedValue({
-      json: async () => ({ version: "2.0.0" }),
+      json: async () => ({
+        announced_version: null,
+        stale: true,
+        version: "2.0.0",
+      }),
     });
     const { findByText } = render(<TestComponent interval={1000} />);
     await act(async () => {
       jest.runOnlyPendingTimers();
     });
     expect(await findByText("stale")).toBeInTheDocument();
+  });
+
+  it("shows stale when the announced version differs from the current bundle", async () => {
+    const { publicEnv } = require("@/config/env");
+    publicEnv.VERSION = "1.0.0";
+    publicEnv.ANNOUNCED_VERSION_ENDPOINT =
+      "https://dnclu2fna0b2b.cloudfront.net/web_build/current-production-version.json";
+    (globalThis.fetch as jest.Mock).mockResolvedValue({
+      json: async () => ({
+        announced_version: "2.0.0",
+        stale: false,
+        version: "2.0.0",
+      }),
+    });
+    const { findByText } = render(<TestComponent interval={1000} />);
+    await act(async () => {
+      jest.runOnlyPendingTimers();
+    });
+    expect(await findByText("stale")).toBeInTheDocument();
+  });
+
+  it("does not use instance version while waiting for a configured announcement", async () => {
+    const { publicEnv } = require("@/config/env");
+    publicEnv.VERSION = "1.0.0";
+    publicEnv.ANNOUNCED_VERSION_ENDPOINT =
+      "https://dnclu2fna0b2b.cloudfront.net/web_build/current-production-version.json";
+    (globalThis.fetch as jest.Mock).mockResolvedValue({
+      json: async () => ({
+        announced_version: null,
+        stale: true,
+        version: "2.0.0",
+      }),
+    });
+    const { findByText } = render(<TestComponent interval={1000} />);
+    await act(async () => {
+      jest.runOnlyPendingTimers();
+    });
+    expect(await findByText("fresh")).toBeInTheDocument();
   });
 
   it("ignores malformed version responses", async () => {
